requirements.txt
numpy
opencv-python
pillow
pyautogui
pytesseract